+++ gameSource/common.h
@@ -12,7 +12,7 @@ typedef struct intPair {
     } intPair;
 
 
-unsigned long intSqrt( unsigned long inX );
+unsigned int intSqrt( unsigned int inX );
 
 // computes angle in degrees [0,360] for a vector from (0,0) to (x,y)
 // rough integer approximation (off by at most 6 degrees for x or y values in
